requirements.txt
cycler==0.10.0
functools32==3.2.3.post2
itk==4.12.0.post1
itk-core==4.12.0.post1
itk-filtering==4.12.0.post1
itk-io==4.12.0.post1
itk-numerics==4.12.0.post1
itk-registration==4.12.0.post1
itk-segmentation==4.12.0.post1
matplotlib==2.0.2
numpy==1.13.1
opencv-python==3.3.0.10
pydicom==0.9.9
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
subprocess32==3.2.7
